Generic programming makes a programmer productive!
 - Type checking done by the compiler! Code is verified at compile time!
 - Data types are reusable!
 - Fun and profit!

Well, not in Java. Java Generics are broken beyond repair. It disturbs everyone, but it doesn't get fixed.

See the AIDS in the diff, to see what it takes to make this simple problem generic!
Also notice the lot of casts and hacks needed to make this simple task compile and work.
Also note: it took a lot of trial and error, because of the casts only fail at runtime! So the main advantage os generics is not present in Java. Just use Object! 

This is when I can understand why the Go folks have not implemented Generics:
If you know cannot (or don't want to) do it properly, you better not even try.
(By cannot I can imagine for example deadlines, budget problem, or other issues related to other engineering decisions)